[[ Linux64 ]] Some more 64-bit cleanliness fixes

+++ engine/src/externalv0.cpp
@@ -1575,8 +1575,8 @@ static char *set_array_utf8_binary(const char *arg1, const char *arg2,
 static char *stack_to_window_rect(const char *arg1, const char *arg2,
 									   const char *arg3, int *retval)
 {
-	uint32_t t_win_id;
-	t_win_id = (uint32_t)arg1;
+	uintptr_t t_win_id;
+	t_win_id = uintptr_t(arg1);
 
 	MCStack *t_stack;
 	t_stack = MCdispatcher->findstackwindowid(t_win_id);
@@ -1603,8 +1603,8 @@ static char *stack_to_window_rect(const char *arg1, const char *arg2,
 static char *window_to_stack_rect(const char *arg1, const char *arg2,
 									   const char *arg3, int *retval)
 {
-	uint32_t t_win_id;
-	t_win_id = (uint32_t)arg1;
+	uintptr_t t_win_id;
+	t_win_id = (uintptr_t)arg1;
 
 	MCStack *t_stack;
 	t_stack = MCdispatcher->findstackwindowid(t_win_id);
